+++ packages/werkbank/lib/src/addons/src/state/src/_internal/immutable/immutable_state_holders_snapshot.dart
@@ -1,0 +1,10 @@
+import 'package:flutter/material.dart';
+
+@immutable
+class ImmutableStateHolderSnapshot {
+  const ImmutableStateHolderSnapshot({
+    required this.value,
+  });
+
+  final Object? value;
+}