+++ packages/werkbank/lib/src/addons/src/state/src/immutable_state_id.dart
@@ -1,0 +1,2 @@
+/// {@category Provide State Without Knobs}
+extension type ImmutableStateId(String _id) {}