+++ packages/werkbank/lib/src/addons/src/state/src/value_container.dart
@@ -1,0 +1,4 @@
+/// {@category Provide State Without Knobs}
+abstract class ValueContainer<T extends Object> {
+  T get value;
+}